+++ src/main/java/omen44/omens_economy/discordLink/Bot.java
@@ -3,10 +3,6 @@ package omen44.omens_economy.discordLink;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.JDABuilder;
 import net.dv8tion.jda.api.entities.Activity;
-import net.dv8tion.jda.api.entities.Message;
-import net.dv8tion.jda.api.entities.MessageChannel;
-import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
-import net.dv8tion.jda.api.hooks.EventListener;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import net.dv8tion.jda.api.requests.GatewayIntent;
 import omen44.omens_economy.Main;
@@ -18,6 +14,7 @@ public class Bot extends ListenerAdapter {
     public Bot(Main plugin) {
         this.plugin = plugin;
     }
+
     public static void main(String[] args) throws LoginException
     {
         if (args.length < 1) {
@@ -27,14 +24,13 @@ public class Bot extends ListenerAdapter {
         // args[0] should be the token
         // We only need 2 intents in this bot. We only respond to messages in guilds and private channels.
         // All other events will be disabled.
-        EventListener register = new CommandRegister(plugin);
         JDA jda = JDABuilder.createLight(args[0], GatewayIntent.GUILD_MESSAGES, GatewayIntent.DIRECT_MESSAGES)
                 .addEventListeners(new Bot(null))
                 .setActivity(Activity.playing("Type !register"))
                 .build();
 
-        jda.addEventListener(new CommandPing());
-        jda.addEventListener();
+        jda.addEventListener(new PingCommand());
+        jda.addEventListener(new RegisterCommand(plugin));
     }
 }
