+++ src/main/java/omen44/omens_economy/discordLink/PingCommand.java
@@ -5,7 +5,7 @@ import net.dv8tion.jda.api.entities.MessageChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 
-public class CommandPing extends ListenerAdapter {
+public class PingCommand extends ListenerAdapter {
     @Override
     public void onMessageReceived(MessageReceivedEvent event)
     {
